+++ zcash_client_sqlite/src/lib.rs
@@ -32,7 +32,7 @@
 // Catch documentation errors caused by code changes.
 #![deny(rustdoc::broken_intra_doc_links)]
 
-use incrementalmerkletree::{Marking, Position, Retention};
+use incrementalmerkletree::Position;
 use nonempty::NonEmpty;
 use rand::RngCore;
 use secrecy::{ExposeSecret, SecretVec};
@@ -40,7 +40,7 @@ use shardtree::{ShardTree, error::ShardTreeError, store::ShardStore};
 use std::{
     borrow::{Borrow, BorrowMut},
     cmp::{max, min},
-    collections::{BTreeSet, HashMap, HashSet},
+    collections::{HashMap, HashSet},
     convert::AsRef,
     fmt,
     num::NonZeroU32,
@@ -48,7 +48,7 @@ use std::{
     path::Path,
 };
 use subtle::ConditionallySelectable;
-use tracing::{debug, trace, warn};
+use tracing::warn;
 use util::Clock;
 use uuid::Uuid;
 
@@ -62,7 +62,7 @@ use zcash_client_backend::{
         WalletRead, WalletSummary, WalletWrite, Zip32Derivation,
         chain::{BlockSource, ChainState, CommitmentTreeRoot},
         ll::{
-            LowLevelWalletRead, LowLevelWalletWrite, ReceivedSaplingOutput,
+            self, LowLevelWalletRead, LowLevelWalletWrite, ReceivedSaplingOutput,
             wallet::store_decrypted_tx,
         },
         scanning::{ScanPriority, ScanRange},
@@ -100,8 +100,7 @@ use wallet::{
 
 #[cfg(feature = "orchard")]
 use {
-    incrementalmerkletree::frontier::Frontier, shardtree::store::Checkpoint,
-    std::collections::BTreeMap, zcash_client_backend::data_api::ORCHARD_SHARD_HEIGHT,
+    zcash_client_backend::data_api::ORCHARD_SHARD_HEIGHT,
     zcash_client_backend::data_api::ll::ReceivedOrchardOutput,
 };
 
@@ -121,12 +120,6 @@ use {
     zcash_keys::encoding::AddressCodec,
 };
 
-#[cfg(feature = "multicore")]
-use maybe_rayon::{
-    prelude::{IndexedParallelIterator, ParallelIterator},
-    slice::ParallelSliceMut,
-};
-
 #[cfg(any(test, feature = "test-dependencies"))]
 use {
     rusqlite::named_params,
@@ -139,18 +132,6 @@ use crate::wallet::encoding::KeyScope;
 
 #[cfg(any(test, feature = "test-dependencies", not(feature = "orchard")))]
 use zcash_protocol::PoolType;
-
-/// `maybe-rayon` doesn't provide this as a fallback, so we have to.
-#[cfg(not(feature = "multicore"))]
-trait ParallelSliceMut<T> {
-    fn par_chunks_mut(&mut self, chunk_size: usize) -> std::slice::ChunksMut<'_, T>;
-}
-#[cfg(not(feature = "multicore"))]
-impl<T> ParallelSliceMut<T> for [T] {
-    fn par_chunks_mut(&mut self, chunk_size: usize) -> std::slice::ChunksMut<'_, T> {
-        self.chunks_mut(chunk_size)
-    }
-}
 
 #[cfg(feature = "unstable")]
 use {
@@ -207,7 +188,7 @@ pub(crate) const UA_TRANSPARENT: ReceiverRequirement = ReceiverRequirement::Requ
 /// events". Examples of these include:
 /// - Restoring a wallet from a backed-up seed.
 /// - Importing the same viewing key into two different wallet instances.
-#[derive(Debug, Copy, Clone, PartialEq, Eq, Hash, Default, PartialOrd, Ord)]
+#[derive(Debug, Copy, Clone, PartialEq, Eq, Default, PartialOrd, Ord, Hash)]
 #[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
 pub struct AccountUuid(#[cfg_attr(feature = "serde", serde(with = "uuid::serde::compact"))] Uuid);
 
@@ -238,7 +219,7 @@ impl AccountUuid {
 ///
 /// This is an ephemeral value for efficiently and generically working with accounts in a
 /// [`WalletDb`]. To reference accounts in external contexts, use [`AccountUuid`].
-#[derive(Debug, Copy, Clone, PartialEq, Eq, Hash, Default, PartialOrd, Ord)]
+#[derive(Debug, Copy, Clone, PartialEq, Eq, Default, PartialOrd, Ord, Hash)]
 pub(crate) struct AccountRef(i64);
 
 /// This implementation is retained under `#[cfg(test)]` for pre-AccountUuid testing.
@@ -252,7 +233,7 @@ impl ConditionallySelectable for AccountRef {
 }
 
 /// An opaque type for received note identifiers.
-#[derive(Debug, Copy, Clone, PartialEq, Eq, PartialOrd, Ord)]
+#[derive(Debug, Copy, Clone, PartialEq, Eq, PartialOrd, Ord, Hash)]
 pub struct ReceivedNoteId(pub(crate) ShieldedProtocol, pub(crate) i64);
 
 impl fmt::Display for ReceivedNoteId {
@@ -264,7 +245,7 @@ impl fmt::Display for ReceivedNoteId {
 }
 
 /// A newtype wrapper for sqlite primary key values for the utxos table.
-#[derive(Debug, Copy, Clone, PartialEq, Eq)]
+#[derive(Debug, Copy, Clone, PartialEq, Eq, PartialOrd, Ord, Hash)]
 pub struct UtxoId(pub(crate) i64);
 
 /// A newtype wrapper for sqlite primary key values for the transactions table.
@@ -272,7 +253,7 @@ pub struct UtxoId(pub(crate) i64);
 pub struct TxRef(pub(crate) i64);
 
 /// A newtype wrapper for sqlite primary key values for the addresses table.
-#[derive(Debug, Copy, Clone, PartialEq, Eq)]
+#[derive(Debug, Copy, Clone, PartialEq, Eq, PartialOrd, Ord, Hash)]
 struct AddressRef(pub(crate) i64);
 
 /// A wrapper for the SQLite connection to the wallet database, along with a capability to read the
@@ -1381,10 +1362,10 @@ impl<C: BorrowMut<rusqlite::Connection>, P: consensus::Parameters, CL: Clock, R:
     }
 
     fn update_chain_tip(&mut self, tip_height: BlockHeight) -> Result<(), Self::Error> {
-        let tx = self.conn.borrow_mut().transaction()?;
-        wallet::scanning::update_chain_tip(&tx, &self.params, tip_height)?;
-        tx.commit()?;
-        Ok(())
+        self.transactionally(|wdb| {
+            wallet::scanning::update_chain_tip(wdb.conn.0, &wdb.params, tip_height)?;
+            Ok(())
+        })
     }
 
     #[tracing::instrument(skip_all, fields(height = blocks.first().map(|b| u32::from(b.height())), count = blocks.len()))]
@@ -1394,424 +1375,11 @@ impl<C: BorrowMut<rusqlite::Connection>, P: consensus::Parameters, CL: Clock, R:
         from_state: &ChainState,
         blocks: Vec<ScannedBlock<Self::AccountId>>,
     ) -> Result<(), Self::Error> {
-        struct BlockPositions {
-            height: BlockHeight,
-            sapling_start_position: Position,
-            #[cfg(feature = "orchard")]
-            orchard_start_position: Position,
-        }
-
-        if blocks.is_empty() {
-            return Ok(());
-        }
-
         self.transactionally(|wdb| {
-            let initial_block = blocks.first().expect("blocks is known to be nonempty");
-            assert!(from_state.block_height() + 1 == initial_block.height());
-
-            let start_positions = BlockPositions {
-                height: initial_block.height(),
-                sapling_start_position: Position::from(
-                    u64::from(initial_block.sapling().final_tree_size())
-                        - u64::try_from(initial_block.sapling().commitments().len()).unwrap(),
-                ),
-                #[cfg(feature = "orchard")]
-                orchard_start_position: Position::from(
-                    u64::from(initial_block.orchard().final_tree_size())
-                        - u64::try_from(initial_block.orchard().commitments().len()).unwrap(),
-                ),
-            };
-
-            let mut sapling_commitments = vec![];
-            #[cfg(feature = "orchard")]
-            let mut orchard_commitments = vec![];
-            let mut last_scanned_height = None;
-            let mut note_positions = vec![];
-
-            #[cfg(feature = "transparent-inputs")]
-            let mut tx_refs = BTreeSet::new();
-
-            for block in blocks.into_iter() {
-                if last_scanned_height
-                    .iter()
-                    .any(|prev| block.height() != *prev + 1)
-                {
-                    return Err(SqliteClientError::NonSequentialBlocks);
-                }
-
-                // Insert the block into the database.
-                wdb.put_block_meta(
-                    block.height(),
-                    block.block_hash(),
-                    block.block_time(),
-                    block.sapling().final_tree_size(),
-                    block.sapling().commitments().len().try_into().unwrap(),
-                    #[cfg(feature = "orchard")]
-                    block.orchard().final_tree_size(),
-                    #[cfg(feature = "orchard")]
-                    block.orchard().commitments().len().try_into().unwrap(),
-                )?;
-
-                for tx in block.transactions() {
-                    let tx_ref = wdb.put_tx_meta(tx, block.height())?;
-
-                    #[cfg(feature = "transparent-inputs")]
-                    tx_refs.insert(tx_ref);
-
-                    wdb.queue_tx_retrieval(std::iter::once(tx.txid()), None)?;
-
-                    // Mark notes as spent and remove them from the scanning cache
-                    for spend in tx.sapling_spends() {
-                        wdb.mark_sapling_note_spent(spend.nf(), tx_ref)?;
-                    }
-                    #[cfg(feature = "orchard")]
-                    for spend in tx.orchard_spends() {
-                        wdb.mark_orchard_note_spent(spend.nf(), tx_ref)?;
-                    }
-
-                    for output in tx.sapling_outputs() {
-                        // Check whether this note was spent in a later block range that
-                        // we previously scanned.
-                        let spent_in = output
-                            .nf()
-                            .map(|nf| wdb.detect_sapling_spend(nf))
-                            .transpose()?
-                            .flatten();
-
-                        wdb.put_received_sapling_note(
-                            output,
-                            tx_ref,
-                            Some(block.height()),
-                            spent_in,
-                        )?;
-                    }
-                    #[cfg(feature = "orchard")]
-                    for output in tx.orchard_outputs() {
-                        // Check whether this note was spent in a later block range that
-                        // we previously scanned.
-                        let spent_in = output
-                            .nf()
-                            .map(|nf| wdb.detect_orchard_spend(nf))
-                            .transpose()?
-                            .flatten();
-
-                        wdb.put_received_orchard_note(
-                            output,
-                            tx_ref,
-                            Some(block.height()),
-                            spent_in,
-                        )?;
-                    }
-                }
-
-                // Insert the new nullifiers from this block into the nullifier map.
-                wdb.track_block_sapling_nullifiers(
-                    block.height(),
-                    block.sapling().nullifier_map(),
-                )?;
-
-                #[cfg(feature = "orchard")]
-                wdb.track_block_orchard_nullifiers(
-                    block.height(),
-                    block.orchard().nullifier_map(),
-                )?;
-
-                note_positions.extend(block.transactions().iter().flat_map(|wtx| {
-                    let iter = wtx.sapling_outputs().iter().map(|out| {
-                        (
-                            ShieldedProtocol::Sapling,
-                            out.note_commitment_tree_position(),
-                        )
-                    });
-                    #[cfg(feature = "orchard")]
-                    let iter = iter.chain(wtx.orchard_outputs().iter().map(|out| {
-                        (
-                            ShieldedProtocol::Orchard,
-                            out.note_commitment_tree_position(),
-                        )
-                    }));
-
-                    iter
-                }));
-
-                last_scanned_height = Some(block.height());
-                let block_commitments = block.into_commitments();
-                trace!(
-                    "Sapling commitments for {:?}: {:?}",
-                    last_scanned_height,
-                    block_commitments
-                        .sapling
-                        .iter()
-                        .map(|(_, r)| *r)
-                        .collect::<Vec<_>>()
-                );
-                #[cfg(feature = "orchard")]
-                trace!(
-                    "Orchard commitments for {:?}: {:?}",
-                    last_scanned_height,
-                    block_commitments
-                        .orchard
-                        .iter()
-                        .map(|(_, r)| *r)
-                        .collect::<Vec<_>>()
-                );
-
-                sapling_commitments.extend(block_commitments.sapling.into_iter().map(Some));
-                #[cfg(feature = "orchard")]
-                orchard_commitments.extend(block_commitments.orchard.into_iter().map(Some));
-            }
-
-            #[cfg(feature = "transparent-inputs")]
-            for (account_id, key_scope) in wdb.find_involved_accounts(tx_refs)? {
-                if let Some(t_key_scope) = key_scope {
-                    use ReceiverRequirement::*;
-                    wdb.generate_transparent_gap_addresses(
-                        account_id,
-                        t_key_scope,
-                        UnifiedAddressRequest::unsafe_custom(Allow, Allow, Require),
-                    )?;
-                }
-            }
-
-            // Prune the nullifier map of entries we no longer need.
-            wdb.prune_tracked_nullifiers(PRUNING_DEPTH)?;
-
-            // We will have a start position and a last scanned height in all cases where
-            // `blocks` is non-empty.
-            if let Some(last_scanned_height) = last_scanned_height {
-                // Create subtrees from the note commitments in parallel.
-                const CHUNK_SIZE: usize = 1024;
-                let sapling_subtrees = sapling_commitments
-                    .par_chunks_mut(CHUNK_SIZE)
-                    .enumerate()
-                    .filter_map(|(i, chunk)| {
-                        let start =
-                            start_positions.sapling_start_position + (i * CHUNK_SIZE) as u64;
-                        let end = start + chunk.len() as u64;
-
-                        shardtree::LocatedTree::from_iter(
-                            start..end,
-                            SAPLING_SHARD_HEIGHT.into(),
-                            chunk.iter_mut().map(|n| n.take().expect("always Some")),
-                        )
-                    })
-                    .map(|res| (res.subtree, res.checkpoints))
-                    .collect::<Vec<_>>();
-
-                #[cfg(feature = "orchard")]
-                let orchard_subtrees = orchard_commitments
-                    .par_chunks_mut(CHUNK_SIZE)
-                    .enumerate()
-                    .filter_map(|(i, chunk)| {
-                        let start =
-                            start_positions.orchard_start_position + (i * CHUNK_SIZE) as u64;
-                        let end = start + chunk.len() as u64;
-
-                        shardtree::LocatedTree::from_iter(
-                            start..end,
-                            ORCHARD_SHARD_HEIGHT.into(),
-                            chunk.iter_mut().map(|n| n.take().expect("always Some")),
-                        )
-                    })
-                    .map(|res| (res.subtree, res.checkpoints))
-                    .collect::<Vec<_>>();
-
-                // Collect the complete set of Sapling checkpoints
-                #[cfg(feature = "orchard")]
-                let sapling_checkpoint_positions: BTreeMap<BlockHeight, Position> =
-                    sapling_subtrees
-                        .iter()
-                        .flat_map(|(_, checkpoints)| checkpoints.iter())
-                        .map(|(k, v)| (*k, *v))
-                        .collect();
-
-                #[cfg(feature = "orchard")]
-                let orchard_checkpoint_positions: BTreeMap<BlockHeight, Position> =
-                    orchard_subtrees
-                        .iter()
-                        .flat_map(|(_, checkpoints)| checkpoints.iter())
-                        .map(|(k, v)| (*k, *v))
-                        .collect();
-
-                #[cfg(feature = "orchard")]
-                fn ensure_checkpoints<
-                    'a,
-                    H,
-                    I: Iterator<Item = &'a BlockHeight>,
-                    const DEPTH: u8,
-                >(
-                    // An iterator of checkpoints heights for which we wish to ensure that
-                    // checkpoints exists.
-                    ensure_heights: I,
-                    // The map of checkpoint positions from which we will draw note commitment tree
-                    // position information for the newly created checkpoints.
-                    existing_checkpoint_positions: &BTreeMap<BlockHeight, Position>,
-                    // The frontier whose position will be used for an inserted checkpoint when
-                    // there is no preceding checkpoint in existing_checkpoint_positions.
-                    state_final_tree: &Frontier<H, DEPTH>,
-                ) -> Vec<(BlockHeight, Checkpoint)> {
-                    ensure_heights
-                        .flat_map(|ensure_height| {
-                            existing_checkpoint_positions
-                                .range::<BlockHeight, _>(..=*ensure_height)
-                                .last()
-                                .map_or_else(
-                                    || {
-                                        Some((
-                                            *ensure_height,
-                                            state_final_tree
-                                                .value()
-                                                .map_or_else(Checkpoint::tree_empty, |t| {
-                                                    Checkpoint::at_position(t.position())
-                                                }),
-                                        ))
-                                    },
-                                    |(existing_checkpoint_height, position)| {
-                                        if *existing_checkpoint_height < *ensure_height {
-                                            Some((
-                                                *ensure_height,
-                                                Checkpoint::at_position(*position),
-                                            ))
-                                        } else {
-                                            // The checkpoint already exists, so we don't need to
-                                            // do anything.
-                                            None
-                                        }
-                                    },
-                                )
-                                .into_iter()
-                        })
-                        .collect::<Vec<_>>()
-                }
-
-                #[cfg(feature = "orchard")]
-                let (missing_sapling_checkpoints, missing_orchard_checkpoints) = (
-                    ensure_checkpoints(
-                        orchard_checkpoint_positions.keys(),
-                        &sapling_checkpoint_positions,
-                        from_state.final_sapling_tree(),
-                    ),
-                    ensure_checkpoints(
-                        sapling_checkpoint_positions.keys(),
-                        &orchard_checkpoint_positions,
-                        from_state.final_orchard_tree(),
-                    ),
-                );
-
-                // Update the Sapling note commitment tree with all newly read note commitments
-                {
-                    let mut sapling_subtrees_iter = sapling_subtrees.into_iter();
-                    wdb.with_sapling_tree_mut::<_, _, SqliteClientError>(|sapling_tree| {
-                        debug!(
-                            "Sapling initial tree size at {:?}: {:?}",
-                            from_state.block_height(),
-                            from_state.final_sapling_tree().tree_size()
-                        );
-                        // We insert the frontier with `Checkpoint` retention because we need to be
-                        // able to truncate the tree back to this point.
-                        sapling_tree.insert_frontier(
-                            from_state.final_sapling_tree().clone(),
-                            Retention::Checkpoint {
-                                id: from_state.block_height(),
-                                marking: Marking::Reference,
-                            },
-                        )?;
-
-                        for (tree, checkpoints) in &mut sapling_subtrees_iter {
-                            sapling_tree.insert_tree(tree, checkpoints)?;
-                        }
-
-                        // Ensure we have a Sapling checkpoint for each checkpointed Orchard block height.
-                        // We skip all checkpoints below the minimum retained checkpoint in the
-                        // Sapling tree, because branches below this height may be pruned.
-                        #[cfg(feature = "orchard")]
-                        {
-                            let min_checkpoint_height = sapling_tree
-                                .store()
-                                .min_checkpoint_id()
-                                .map_err(ShardTreeError::Storage)?
-                                .expect(
-                                    "At least one checkpoint was inserted (by insert_frontier)",
-                                );
-
-                            for (height, checkpoint) in &missing_sapling_checkpoints {
-                                if *height > min_checkpoint_height {
-                                    sapling_tree
-                                        .store_mut()
-                                        .add_checkpoint(*height, checkpoint.clone())
-                                        .map_err(ShardTreeError::Storage)?;
-                                }
-                            }
-                        }
-
-                        Ok(())
-                    })?;
-                }
-
-                // Update the Orchard note commitment tree with all newly read note commitments
-                #[cfg(feature = "orchard")]
-                {
-                    let mut orchard_subtrees = orchard_subtrees.into_iter();
-                    wdb.with_orchard_tree_mut::<_, _, SqliteClientError>(|orchard_tree| {
-                        debug!(
-                            "Orchard initial tree size at {:?}: {:?}",
-                            from_state.block_height(),
-                            from_state.final_orchard_tree().tree_size()
-                        );
-                        // We insert the frontier with `Checkpoint` retention because we need to be
-                        // able to truncate the tree back to this point.
-                        orchard_tree.insert_frontier(
-                            from_state.final_orchard_tree().clone(),
-                            Retention::Checkpoint {
-                                id: from_state.block_height(),
-                                marking: Marking::Reference,
-                            },
-                        )?;
-
-                        for (tree, checkpoints) in &mut orchard_subtrees {
-                            orchard_tree.insert_tree(tree, checkpoints)?;
-                        }
-
-                        // Ensure we have an Orchard checkpoint for each checkpointed Sapling block height.
-                        // We skip all checkpoints below the minimum retained checkpoint in the
-                        // Orchard tree, because branches below this height may be pruned.
-                        {
-                            let min_checkpoint_height = orchard_tree
-                                .store()
-                                .min_checkpoint_id()
-                                .map_err(ShardTreeError::Storage)?
-                                .expect(
-                                    "At least one checkpoint was inserted (by insert_frontier)",
-                                );
-
-                            for (height, checkpoint) in &missing_orchard_checkpoints {
-                                if *height > min_checkpoint_height {
-                                    debug!(
-                                        "Adding missing Orchard checkpoint for height: {:?}: {:?}",
-                                        height,
-                                        checkpoint.position()
-                                    );
-                                    orchard_tree
-                                        .store_mut()
-                                        .add_checkpoint(*height, checkpoint.clone())
-                                        .map_err(ShardTreeError::Storage)?;
-                                }
-                            }
-                        }
-                        Ok(())
-                    })?;
-                }
-
-                wdb.notify_scan_complete(
-                    Range {
-                        start: start_positions.height,
-                        end: last_scanned_height + 1,
-                    },
-                    &note_positions,
-                )?;
-            }
-
-            Ok(())
+            ll::wallet::put_blocks::<_, SqliteClientError, commitment_tree::Error>(
+                wdb, from_state, blocks,
+            )
+            .map_err(SqliteClientError::from)
         })
     }