+++ zcash_client_sqlite/CHANGELOG.md
@@ -20,6 +20,8 @@ workspace.
 - `impl<'a> Borrow<rusqlite::Transaction<'a>> for zcash_client_sqlite::SqlTransaction<'a>`
 - `impl zcash_client_backend::data_api::ll::LowLevelWalletRead for WalletDb`
 - `impl zcash_client_backend::data_api::ll::LowLevelWalletWrite for WalletDb`
+- `impl Hash for zcash_client_sqlite::{ReceivedNoteId, UtxoId, TxRef}`
+- `impl {PartialOrd, Ord} for zcash_client_sqlite::{UtxoId, PartialOrd, Ord}`
 
 ### Changed
 - Migrated to `orchard 0.12`, `sapling-crypto 0.6`.